requirements.txt
streamlit>=1.24.0
pymupdf==1.24.11
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers>=4.35.0
torch>=2.0.0
numpy
tqdm
typing_extensions
